requirements.txt
cryptography>=41.0
PyYAML>=6.0
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
jupyter>=1.0
